move exceptions and related helpers to scrapinghub.exceptions module
update docs with short exceptions descriptions
