+++ scrapinghub/utils.py
@@ -1,50 +1,6 @@
 import logging
-from functools import wraps
 
 from six import string_types
-from requests.exceptions import HTTPError
-
-from scrapinghub.hubstorage import ValueTooLarge as _ValueTooLarge
-
-
-class ScrapinghubAPIError(Exception):
-
-    def __init__(self, message=None, http_error=None):
-        self.http_error = http_error
-        if not message:
-            message = _get_http_error_msg(http_error)
-        super(ScrapinghubAPIError, self).__init__(message)
-
-
-def _get_http_error_msg(exc):
-    if isinstance(exc, HTTPError):
-        try:
-            error_obj = exc.response.json()
-        except ValueError:
-            error_obj = None
-        if error_obj and isinstance(error_obj, dict):
-            error = error_obj.get('error')
-            if error:
-                return error
-        elif exc.response.text:
-                return exc.response.text
-    return str(exc)
-
-
-class InvalidUsage(ScrapinghubAPIError):
-    pass
-
-
-class NotFound(ScrapinghubAPIError):
-    pass
-
-
-class ValueTooLarge(ScrapinghubAPIError):
-    pass
-
-
-class DuplicateJobError(ScrapinghubAPIError):
-    pass
 
 
 class LogLevel(object):
@@ -103,40 +59,11 @@ def get_tags_for_update(**kwargs):
     return params
 
 
-def wrap_http_errors(method):
-    @wraps(method)
-    def wrapped(*args, **kwargs):
-        try:
-            return method(*args, **kwargs)
-        except HTTPError as exc:
-            status_code = exc.response.status_code
-            if status_code == 400:
-                raise InvalidUsage(http_error=exc)
-            elif status_code == 404:
-                raise NotFound(http_error=exc)
-            elif status_code == 413:
-                raise ValueTooLarge(http_error=exc)
-            elif status_code > 400 and status_code < 500:
-                raise ScrapinghubAPIError(http_error=exc)
-            raise
-    return wrapped
-
-
 def wrap_kwargs(fn, kwargs_fn):
     """Tiny wrapper to prepare modified version of function kwargs"""
     def wrapped(*args, **kwargs):
         kwargs = kwargs_fn(kwargs)
         return fn(*args, **kwargs)
-    return wrapped
-
-
-def wrap_value_too_large(method):
-    @wraps(method)
-    def wrapped(*args, **kwargs):
-        try:
-            return method(*args, **kwargs)
-        except _ValueTooLarge as exc:
-            raise ValueTooLarge(str(exc))
     return wrapped
